requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\Hossein\diabetes_detection\webapp.py: 19
Pillow == 7.2.0

# C:\Users\Hossein\diabetes_detection\webapp.py: 15
pandas == 1.0.5

# C:\Users\Hossein\diabetes_detection\webapp.py: 16,17,18
scikit_learn == 0.23.1

# C:\Users\Hossein\diabetes_detection\webapp.py: 20
streamlit == 0.72.0
